test indices to ints

+++ cebra_embedding.py
@@ -31,7 +31,7 @@ def create_folds(n_timesteps, num_folds=5, num_windows=10):
         test_windows = np.arange(i, n_windows_total, num_folds)
         test_ind = []
         for j in test_windows:
-            test_ind.extend(np.arange(window_start_ind[j], window_start_ind[j] + np.round(window_size)))
+            test_ind.extend(np.int32(np.arange(window_start_ind[j], window_start_ind[j] + np.round(window_size))))
 
         # remove any test_ind that greater than n_timesteps-1
         test_ind = [ind for ind in test_ind if ind < n_timesteps]
